requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
python-dotenv
pyodbc
reportlab
SQLAlchemy
